chore(api): extracted the reading API contracts to re-export as shared types
